# api/src/openapi.yml
openapi: 3.0.3
info:
  title: Offline-First Chat Application API
  description: REST API for an offline-first chat application with event-based synchronization and configurable conflict resolution
  version: 1.0.0
servers:
  - url: https://api.chatapp.com/v1
    description: Production server
paths:
...
  /admin/conflictPolicies:
    get:
      summary: Get all conflict resolution policies (admin only)
      tags: [Admin]
      responses:
        '200':
          description: List of conflict resolution policies
          content:
            application/json:
              schema:
                type: object
                properties:
                  policies:
                    type: array
                    items:
                      $ref: '#/components/schemas/ConflictPolicy'

    post:
      summary: Create or update a conflict resolution policy (admin only)
      tags: [Admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConflictPolicyInput'
      responses:
        '201':
          description: Policy created/updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictPolicy'
  /admin/conflictPolicies/{policyId}:
    delete:
      summary: Delete a conflict resolution policy (admin only)
      tags: [Admin]
      parameters:
        - name: policyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Policy deleted successfully

  # Users (basic endpoints for hardcoded users)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AdminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin-level JWT token required
  schemas:
...
    ConflictPolicy:
      type: object
      properties:
        id:
          type: string
          format: uuid
        policyName:
          type: string
        entityType:
          type: string
          enum: [MESSAGE, MEMBERSHIP, GROUP]
        conflictType:
          type: string
        resolutionStrategy:
          type: string
        parameters:
          type: object
          additionalProperties: true
        isActive:
          type: boolean
        createdByAdminId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    ConflictPolicyInput:
      type: object
      properties:
        policyName:
          type: string
          maxLength: 100
        entityType:
          type: string
          enum: [MESSAGE, MEMBERSHIP, GROUP]
        conflictType:
          type: string
          maxLength: 100
        resolutionStrategy:
          type: string
          maxLength: 100
        parameters:
          type: object
          additionalProperties: true
        isActive:
          type: boolean
          default: true
      required: [policy_name, entity_type, conflict_type, resolution_strategy]
...